Remove integer typedef

+++ docs/old_code/tcc/compiler_old/ast.cpp
@@ -3,11 +3,11 @@
 namespace tcc
 {
 
-Integer AppendExpression(InstructionList& dest, Expression const& source)
+int64_t AppendExpression(InstructionList& dest, Expression const& source)
 {
     auto const assembly = source.GetAssembly();
     for (auto const& x : assembly) dest.push_back(x);
 
-    return static_cast<Integer>(assembly.size());
+    return static_cast<int64_t>(assembly.size());
 }
 }  // namespace tcc
